refactor: moved the map to staking.rs
